t4210: correctly check for invalid UTF-8 in git log --grep
7187c7bbb8 (t4210: skip i18n tests that don't work on FreeBSD, 2019-11-27)
adds a REG_ILLSEQ prerequisite to avoid failures from the tests added with
4e2443b181 (log tests: test regex backends in "--encode=<enc>" tests,
2019-06-28)

Instead of disabling the problematic tests, make sure that their definition
and implementation are synced, and that the intention was to replicate
all negative tests with all valid permutations (based on the original
descriptions).

While at it remove the prerequisites and make sure that the behaviour
that is expected from POSIX regcomp[1] is tested against; meaning that
an error is also a valid response for invalid encoding, as far as no
abort is triggered.

Note that while not technicaly valid, I am keeping the check for the empty
result, as it will be very unlikely that anything with that needle will
match in the currently provided stash, but "unspecified" behaviour doesn't
preclude it.

[1] https://pubs.opengroup.org/onlinepubs/9699919799/functions/regcomp.html
and make sure