win32: avoid w_char strings from standard error functions
f7adba4182 (cmake: support for building git on windows with mingw,
2020-06-26) enables UNICODE for VS builds (really in CMake), but
didn't prevent Windows to alias some functions that are otherwise
always expected to return "char *" into their wide versions.

gai_strerror() is used to print an error message in the console
when using imap-send (as well as other), so make sure that function
is aliased to the ANSI version so it matches the "%s" format string
that will be used.

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>