t/helper/test-drop-caches: avoid UNICODE for compatibility
14527b3002 (fsmonitor: add a performance test, 2017-09-22), adds
code to drop the file cache that used 2 functions that could be
aliased when UNICODE is enabled.

Make sure that they use the "A" alias to avoid changing behaviour
or the buffer types they use and to keep them working as they did
before UNICODE was enabled (currently only in VS builds).

Alternatively, the code could be transformed to support UNICODE
(if watchman or any other related tool like FSMonitor does), but
that has been punted for now, as it will probably also might need
changes to support long paths anyway.

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>