tr2: support both ANSI and UNICODE for process names in Windows
353d3d77f4 (trace2: collect Windows-specific process information,
2019-02-22), add a tracer for executable names that uses a struct
that could be aliased in UNICODE mode to use wchar_t instead.

Add a branch that supports UNICODE and transforms the executable
name into UTF-8 before adding it to the JSON, but keep the
original implementation which should be better performing and is
used by the usual instalation.

Alternatively, the ANSI code could be dropped, or it could be made
ANSI dependent and drop the UNICODE one, if executable names with
non ASCII names aren't supported.

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>