terminal: teach git how to save/restore its terminal settings
Add two functions to push/pop the terminal settings using either
the POSIX or Windows console functions, and refactor the current
code to use them.

This will allow for the state of the terminal to be saved and
restored around a child that might need to change them (ex vi)
and might not restore them properly upon exit.

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>