terminal: teach git_terminal_prompt() to fail if background
21aeafceda (add generic terminal prompt function, 2011-12-10) add
a custom replacement function for getpass() that could be used to
get user input in a better way, but the behaviour when not
foreground was not implemented, leading for it to write its prompt
into the console, even when not in the foreground (if allowed by
`stty -tostop` which seems to be the default).

Replace the original check with a new one that will fail early if
not foreground, and this way prevent any writes or the blocking of
its process through SIGTTIN because of later background reads
(when echo was enabled).

After this change the following command when run in the background
(and assuming there is no configured credential for that url) will
not write anything to the console or get stopped, while failing
with a more useful error than before.

  ( echo "url=https://example.com/" | git credential fill ) &

The original check for when echo is disabled is done earlier, so
it will still get blocked when asking for a Password just like it
did before, and like getpass() does, but unlike getpass() it won't
be able to recover if foreground, so that will be changed next for
consistency.

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>