terminal: fail early if not foreground regardless of mode
Since 21aeafceda (add generic terminal prompt function, 2011-12-10)
platforms using HAVE_DEV_TTY=1 (which are most of them) are using
our own implementation of a getpass() like function that is missing
support for intercepting SIGTTOU/SIGTTIN and recovering after
SIGSTOP, but when echo is suppressed (for the most common setup) it
will still trigger a SIGTTOU that git won't be able to recover from.

Remove the possibility of triggering that by detecting if running in
the background and failing early, and therefore make consistent the
failure mode that was introduced in a previous commit for all calls
to this function when in the background.

While this is a change on behaviour from the current state, it is
expected to be beneficial as it will prevent git getting stopped
accidentally if this code is ever called by a daemon like process,
and since the function wouldn't be able to recover anyway, makes the
failure mode simpler even if it also differs from getpass().

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>