terminal: supress stty execution errors in disable_bits()
9ea416cb51 (terminal: accommodate Git for Windows' default terminal,
2020-01-14) allow Git for Windows to invoke MSYS2 stty to configure
the terminal, but that program might not be always accessible or in
PATH.

Suppress any errors and continue silently to the available fallback.

While at it, move the fallback for macro for FORCE_TEXT that was
introduced in afb43561b8 (mingw: reuse tty-version of
git_terminal_prompt, 2012-12-04) together with the other 2 macros
that are required for each implementation for easy of maintenance.

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>