git-daemon: use sigaction() to install child_handler()
Since bd7b371e9c (git-daemon: SysV needs the signal handler reinstated.,
2008-08-12), the child_handler() has been calling signal() to rearm its
handler, but that crashes in at least AIX.

signal() needs the rearm in SysV but does not in BSD, and which one is
being used can be confusing. POSIX provides and recommends sigaction()
as an alternative because of that, so use that instead.

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>