daemon: introduce ppoll() to the main loop
`service_loop()` uses poll() to keep track of all the listener
sockets, and relies on it being interrupted by SIGCHLD to keep
the number of zombie processes on check.

That last part has become more unreliable since SA_RESTART was
invented and as it has become more commonly invoked.

`ppoll()` was introduced with IEEE Std 1003.1-2024 to POSIX
as an extension of `poll()`, that includes a list of signal
handlers that are not blocked during its execution, and which
will trigger the function to return -1 in those cases, with
errno set to EINTR.

Add the alternative calls, behind a HAVE_PPOLL Makefile flag
and enable it for recent versions of OpenBSD and NetBSD.

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>

[1] https://pubs.opengroup.org/onlinepubs/9799919799/