daemon: add a self pipe to trigger reaping of children
Since 695605b5080 (git-daemon: Simplify dead-children reaping logic,
2008-08-14), there is a race condition between the time a children
status is checked, and the arrival of a SIGCHLD that would alert us
of their demise, hopefully by interrupting poll().

To close the gap, add the reading side of a pipe to the poll and
use the signal handler to write to it for each invocation, so if
the SIGCHLD is somehow missed, poll() would still be interrupted.

Additionally, add an accumulator to track the number of invocations
that could be used even when the pipe isn't available.

Suggested-by: Phillip Wood <phillip.wood123@gmail.com>
Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>