wrapper: handle EINTR from internall poll() in x[read|write]
Since 1079c4be0b (xread: poll on non blocking fds, 2015-12-15)
a poll() was introduced to xread to wait when a O_NONBLOCK fd
was being used, similar logic was later added for xwrite and
eventually factored out, while retaining the same logic.

An EINTR in poll() doesn't indicate a change on the status of
the fd that is being read/written to, and therefore it should
be handled internally.

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>