compat/terminal: remove use of msys2 stty
Introduced with 9ea416cb51 (terminal: accommodate Git for Windows'
default terminal, 2020-01-14), to workaround the fact that git is
not a msys2 application, but sometime later it started to fail[1]
silently.

Remove the implementation, and let the code fallback to the native
code which works both with and without mintty since the later uses
conpty by default for its tty implementation.

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>

[1] https://github.com/msys2/MSYS2-packages/issues/5498