compat/terminal: use binary to read in git_terminal_prompt()
Since afb43561b8 (mingw: reuse tty-version of git_terminal_prompt,
2012-12-04), the Windows version, opens the terminal in text mode
to avoid having to deal with the different EOL, but that has the
side effect of forcing any text thah contain non ASCII code to be
translated to whatever CP the terminal has, mangling UTF-8.

Ideally, the CP, msys2 locale and stty iflags should be all in
sync, but solving that is out of scope for this simple change
which could be addressed by simply avoiding any text translation.

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>